Disable Cxx in REQUIRE to support system-wide Cxx.jl installations
Cxx is a heavyweight package and can take a long time to build, so it can
be very beneficial to install it system-wide (in a "share/julia/site"
directory) - especially as ROOT.jl currently requires a modified version
of Cxx.jl for RTTI support. However, if Cxx.jl is installed in `LOAD_PATH`
instead of `Pkg.dir()`, it doesn't satisfy a REQUIRE entry. Hopefully
Pkg3 will improve the situation.

Also:

* Require Julia 0.5 stable release.
* Don't require BinDeps as we currently don't use it.