Full reimplementation using CxxWrap instead of Cxx.
This dramatic change was required to work with Julia releases. Previous code could not work with releases > 1.3. Contrary to Cxx, which was providing dynamic access to the full C++ API, CxxWrap requires writing C++ wrapping code for each type and function/method to bind. Wrapper code has been written, thanks to automatic code generation, for a number of ROOT C++ classes and their methods. More work is needed to cover the full C++ API.

The code is essentially the one from the ex002-ROOT example of wrapit (release 1.3.1) packages. Mechanism to allow installation from the package manager, including installation of the ROOT C++ libraries was added. The ROOT C++ libraries are currently not available as a _jll package due to difficulties encountered with the cross-compilation required by the Julia binary distribution system.